Refactor code for adding and showing people
